feature(core): product group placeholder
