+++ modules/@ergonode/products/src/extends/components/Buttons/AddProductsToGroupButton.vue
@@ -3,87 +3,38 @@
  * See LICENSE for license details.
  */
 <template>
-    <ActionButton
+    <Button
         title="ADD PRODUCTS"
-        :theme="secondaryTheme"
         :disabled="!isAllowedToUpdate"
         :size="smallSize"
-        :options="addProductOptions"
-        :fixed-content="true"
-        @input="onSelectAddProductOption">
-        <template #prepend="{ color }">
-            <IconAdd :fill-color="color" />
-        </template>
-        <Component
-            v-if="selectedModal"
-            :is="modalComponent"
+        @click.native="onShowModal">
+        <AddProductsToGroupModalGrid
+            v-if="isModalVisible"
             @close="onCloseModal" />
-    </ActionButton>
+    </Button>
 </template>
 
 <script>
 import {
     SIZE,
-    THEME,
 } from '@Core/defaults/theme';
 import PRIVILEGES from '@Products/config/privileges';
-import {
-    ADD_PRODUCT,
-} from '@Products/defaults';
-import ActionButton from '@UI/components/ActionButton/ActionButton';
-import IconAdd from '@UI/components/Icons/Actions/IconAdd';
+import Button from '@UI/components/Button/Button';
 
 export default {
     name: 'AddProductsToGroupButton',
     components: {
-        IconAdd,
-        ActionButton,
+        Button,
+        AddProductsToGroupModalGrid: () => import('@Products/extends/components/Modals/AddProductsToGroupModalGrid'),
     },
     data() {
         return {
-            selectedModal: null,
+            isModalVisible: false,
         };
     },
     computed: {
         smallSize() {
             return SIZE.SMALL;
-        },
-        secondaryTheme() {
-            return THEME.SECONDARY;
-        },
-        addProductOptions() {
-            const options = Object.values(ADD_PRODUCT);
-
-            if (this.extendedComponents.length) {
-                this.extendedComponents.forEach((option) => {
-                    options.push(option.name);
-                });
-            }
-
-            return options;
-        },
-        extendedComponents() {
-            return this.$getExtendSlot('@Products/components/Tabs/ProductGroupTab/addProductFrom') || [];
-        },
-        modalComponent() {
-            let extendedOptions = [];
-
-            if (this.extendedComponents.length) {
-                extendedOptions = this.extendedComponents;
-            }
-            const modals = [
-                {
-                    component: () => import('@Products/extends/components/Modals/AddProductsToGroupModalGrid'),
-                    name: ADD_PRODUCT.FROM_LIST,
-                },
-                {
-                    component: () => import('@Products/extends/components/Modals/AddProductsFromSegmentModalForm'),
-                    name: ADD_PRODUCT.FROM_SEGMENT,
-                },
-                ...extendedOptions,
-            ];
-
-            return modals.find(modal => modal.name === this.selectedModal).component;
         },
         isAllowedToUpdate() {
             return this.$hasAccess([
@@ -92,11 +43,11 @@ export default {
         },
     },
     methods: {
-        onSelectAddProductOption(option) {
-            this.selectedModal = option;
+        onShowModal() {
+            this.isModalVisible = true;
         },
         onCloseModal() {
-            this.selectedModal = null;
+            this.isModalVisible = false;
         },
     },
 };
