+++ modules/@ergonode/products/src/locales/en_GB.json
@@ -9,7 +9,13 @@
         },
         "grid": {
             "placeholderTitle": "No products",
-            "placeholderSubtitle": "There are no product in the system, you can create the first one."
+            "placeholderSubtitle": "There are no product in the system, you can create the first one.",
+            "groupPlaceholderTitle": "No products in group",
+            "groupPlaceholderSubtitle": "There are no product added to the group, you can add the first one.",
+            "variantPlaceholderTitle": "No product variants",
+            "variantPlaceholderSubtitle": "Choose products which together will create a product with variants. You can choose between products which contain selected binding attributes.",
+            "productsToAttachForGroupPlaceholderTitle": "No products",
+            "productsToAttachForGroupPlaceholderSubtitle": "There are either no product with variants nor simple products in the system, you can create the first one."
         }
     }
 }
