+++ modules/@ergonode/products/src/locales/pl_PL.json
@@ -9,7 +9,13 @@
         },
         "grid": {
             "placeholderTitle": "Brak produktów",
-            "placeholderSubtitle": "W systemie nie ma produktów, stwórz pierwszy."
+            "placeholderSubtitle": "W systemie nie ma produktów, stwórz pierwszy.",
+            "groupPlaceholderTitle": "Brak produktów w grupie",
+            "groupPlaceholderSubtitle": "Brak produktów w grupie, dodaj pierwszy.",
+            "variantPlaceholderTitle": "Brak wariantów produktowych",
+            "variantPlaceholderSubtitle": "Wybierz produkty, które wspólnie stworzą produkt z wariantami. Możesz wybrać między produktami, które posiadają wartości z dowiązanych atrybutów.",
+            "productsToAttachForGroupPlaceholderTitle": "Brak produktów",
+            "productsToAttachForGroupPlaceholderSubtitle": "Brak produktów z wariantami albo produktów prostych w systemie, stwórz pierwszy."
         }
     }
 }
